+++ src/main/java/com/publicarttrail/trails/entities/Trail.java
@@ -7,17 +7,16 @@ import lombok.RequiredArgsConstructor;
 import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 
 @Data // Creates all getters, setters, etc. for all attributes
 @RequiredArgsConstructor
 @Entity(name = "Trail")
 @Table(name = "trail")
 public class Trail {
-    @Id                                                 // Indicate that this is the primary key of the table
-    @GeneratedValue(strategy = GenerationType.IDENTITY) // Auto-increment in the table (We don't need to provide a table as it's automatically generated for us)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private long id;                                    // Primary key
+    private int id;
 
     @Column(name = "name")
     private String name;
